==> Huang_PDOs/2_cycle_annotation.Rmd <==
---
title: "2.Huang_2023 (PDOs): cell cycle assignation"
author: "Romuald Parmentier"
date: "2024-03-26"
output: html_document
---

```{r Load libraries and create output folder, message = F}

# Libraries
library(SingleCellExperiment)
library(Seurat)
library(ggplot2)
library(dplyr)

# Functions and palettes
source(paste0(github_dir, "/Dolgos_Custom_Functions.R"))
source(paste0(github_dir, "/Dolgos_Custom_Color_Palettes.R"))

# Create output path
out_path = create_exp_folder(
  github_dir = github_dir,
  samples_ID = "Huang_PDOs",
  exp = "2_cycle_annotation"
)

```


```{r Load files}

# Load sce_comb after normalization
file_path = get_exp_file_path(
  github_dir = github_dir,
  samples_ID = "Huang_PDOs",
  prev_exp = "1_normalization_rescaling", 
  pattern = "sce_comb_normalized_rescaled.rds")

sce_comb = readRDS(file = file_path) 

```

# Cell type assignation 

Already present in the downloaded object

# Cell cycle assignation 

```{r Cell cycle phase prediction}

# Transform sce object in seurat object
seurat_comb = as.Seurat(
  x = sce_comb,
  counts = "counts",
  data = NULL
)

# Normalize and scale the seurat object according standard Seurat procedure
DefaultAssay(object = seurat_comb) = "originalexp"
seurat_comb <- NormalizeData(seurat_comb, normalization.method = "LogNormalize", scale.factor = 10000)
seurat_comb <- ScaleData(seurat_comb)

# Store cell cycles-associated genes in variables
s.genes <- cc.genes.updated.2019$s.genes
g2m.genes <- cc.genes.updated.2019$g2m.genes

# Compute the cell cycle signature score
seurat_comb <- CellCycleScoring(seurat_comb, s.features = s.genes, g2m.features = g2m.genes, set.ident = TRUE)

# Add cell cycle phase to the sce object
sce_comb$cell_cycle_phase = seurat_comb$Phase

# Print cell cycles phases stats
table(sce_comb$cell_cycle_phase)

```


# Export files

```{r}

# Save annotated sce object
saveRDS(
  object = sce_comb,
  file = paste0(out_path,time_stamp(),"sce_comb_cycle_annotated.rds")
  )


```
